refactor: migrate repositories and services into features
