+++ packages/@foodbudget/app/src/index.ts
@@ -1,3 +1,3 @@
 export { default as serviceManager, ServiceManager } from './serviceManager';
-export { default as Repository } from './repository';
-export { Recipe } from './repository/recipe';
+export { default as Repository } from './types';
+export { Recipe } from './recipes/repository';
