+++ packages/@foodbudget/app/src/serviceManager/serviceManager.types.ts
@@ -1,4 +1,4 @@
-import { RecipeServicesInterface } from '../services/recipe';
+import { RecipeServicesInterface } from '../recipes/services';
 
 export interface ServiceManager {
     recipeServices: RecipeServicesInterface;
